requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
absl-py==0.6.1            # via tensorflow
astor==0.7.1              # via tensorflow
gast==0.2.0               # via tensorflow
grpcio==1.16.0            # via tensorboard, tensorflow
h5py==2.8.0               # via keras, keras-applications
imageio==2.4.1
keras-applications==1.0.6  # via keras, tensorflow
keras-preprocessing==1.0.5  # via keras, tensorflow
keras==2.2.4
markdown==3.0.1           # via tensorboard
numpy==1.15.3             # via h5py, imageio, keras, keras-applications, keras-preprocessing, scipy, tensorboard, tensorflow
pillow==5.3.0
protobuf==3.6.1           # via tensorboard, tensorflow
pyyaml==3.13              # via keras
scipy==1.1.0
six==1.11.0               # via absl-py, grpcio, h5py, keras, keras-preprocessing, protobuf, tensorboard, tensorflow
tensorboard==1.11.0       # via tensorflow
tensorflow==1.11.0
termcolor==1.1.0          # via tensorflow
tqdm==4.28.1
werkzeug==0.14.1          # via tensorboard
wheel==0.32.2             # via tensorboard, tensorflow
